requirements.txt
g4f~=0.3.8.0
setuptools
wheel
requests==2.32.0
moviepy~=2.1.1
termcolor==2.4.0
flask==3.0.0
curl_cffi
flask-cors==4.0.0
playsound==1.3.0
Pillow==9.5.0
python-dotenv==1.0.0
srt_equalizer==0.1.8
platformdirs==4.1.0
undetected_chromedriver
assemblyai~=0.35.1
brotli
google-api-python-client~=2.154.0
oauth2client~=4.1.3
openai~=1.55.3
google-generativeai

protobuf~=5.29.0
numpy~=2.1.3
httplib2~=0.22.0